+++ sdk/monitor/azquery/CHANGELOG.md
@@ -9,6 +9,10 @@
 
 ### Breaking Changes
 * Added error return values to `NewLogsClient` and `NewMetricsClient`
+* Rename `Batch` to `QueryBatch`
+* Rename `NewListMetricDefinitionsPager` to `NewListDefinitionsPager`
+* Rename `NewListMetricNamespacesPager` to `NewListNamespacesPager`
+* Changed type of `Render` and `Statistics` from interface{} to []byte
 
 ### Bugs Fixed